refactor :  breaking the util into helper files
