+++ docetl/helper/api_wrapper.py
@@ -1,381 +1,18 @@
 import ast
-import functools
 import hashlib
 import json
-import os
-import shutil
-import threading
-from concurrent.futures import as_completed
-from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple, Union
-
-import litellm
-import tiktoken
-from asteval import Interpreter
-from diskcache import Cache
-from dotenv import load_dotenv
-from frozendict import frozendict
-from jinja2 import Template
-from litellm import completion, embedding, model_cost, RateLimitError
-from rich import print as rprint
-from rich.console import Console
-from rich.prompt import Prompt
-from tqdm import tqdm
-from pydantic import BaseModel
-
-from docetl.console import DOCETL_CONSOLE
-from docetl.utils import completion_cost, count_tokens
 import time
-from litellm.utils import ModelResponse
+from typing import List, Dict, Optional, Any
 
-aeval = Interpreter()
+from jinja2 import Template
+from litellm import embedding, completion, RateLimitError
+from litellm.types.utils import ModelResponse
+from rich import print as rprint, Console
 
-load_dotenv()
-# litellm.set_verbose = True
-DOCETL_HOME_DIR = os.path.expanduser("~/.docetl")
-
-CACHE_DIR = os.path.join(DOCETL_HOME_DIR, "cache")
-LLM_CACHE_DIR = os.path.join(DOCETL_HOME_DIR, "llm_cache")
-cache = Cache(LLM_CACHE_DIR)
-cache.close()
-
-
-class LLMResult(BaseModel):
-    response: Any
-    total_cost: float
-    validated: bool
-
-
-def freezeargs(func):
-    """
-    Decorator to convert mutable dictionary arguments into immutable.
-
-    This decorator is useful for making functions compatible with caching mechanisms
-    that require immutable arguments.
-
-    Args:
-        func (callable): The function to be wrapped.
-
-    Returns:
-        callable: The wrapped function with immutable dictionary arguments.
-    """
-
-    @functools.wraps(func)
-    def wrapped(*args, **kwargs):
-        args = tuple(
-            (
-                frozendict(arg)
-                if isinstance(arg, dict)
-                else json.dumps(arg) if isinstance(arg, list) else arg
-            )
-            for arg in args
-        )
-        kwargs = {
-            k: (
-                frozendict(v)
-                if isinstance(v, dict)
-                else json.dumps(v) if isinstance(v, list) else v
-            )
-            for k, v in kwargs.items()
-        }
-        return func(*args, **kwargs)
-
-    return wrapped
-
-
-def flush_cache(console: Console = DOCETL_CONSOLE):
-    """
-    Flush the cache to disk.
-    """
-    console.log("[bold green]Flushing cache to disk...[/bold green]")
-    cache.close()
-    console.log("[bold green]Cache flushed to disk.[/bold green]")
-
-
-def clear_cache(console: Console = DOCETL_CONSOLE):
-    """
-    Clear the LLM cache stored on disk.
-
-    This function removes all cached items from the disk-based cache,
-    effectively clearing the LLM's response history.
-
-    Args:
-        console (Console, optional): A Rich console object for logging.
-            Defaults to a new Console instance.
-    """
-    console.log("[bold yellow]Clearing LLM cache...[/bold yellow]")
-    try:
-        with cache as c:
-            c.clear()
-        # Remove all files in the cache directory
-        cache_dir = CACHE_DIR
-        if not os.path.exists(cache_dir):
-            os.makedirs(cache_dir)
-        for filename in os.listdir(cache_dir):
-            file_path = os.path.join(cache_dir, filename)
-            try:
-                if os.path.isfile(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                console.log(
-                    f"[bold red]Error deleting {file_path}: {str(e)}[/bold red]"
-                )
-        console.log("[bold green]Cache cleared successfully.[/bold green]")
-    except Exception as e:
-        console.log(f"[bold red]Error clearing cache: {str(e)}[/bold red]")
-
-def convert_dict_schema_to_list_schema(schema: Dict[str, Any]) -> Dict[str, Any]:
-    schema_str = "{" + ", ".join([f"{k}: {v}" for k, v in schema.items()]) + "}"
-    return {"results": f"list[{schema_str}]"}
-
-def convert_val(value: Any, model: str = "gpt-4o-mini") -> Dict[str, Any]:
-    """
-    Convert a string representation of a type to a dictionary representation.
-
-    This function takes a string value representing a data type and converts it
-    into a dictionary format suitable for JSON schema.
-
-    Args:
-        value (Any): A string representing a data type.
-        model (str): The model being used. Defaults to "gpt-4o-mini".
-
-    Returns:
-        Dict[str, Any]: A dictionary representing the type in JSON schema format.
-
-    Raises:
-        ValueError: If the input value is not a supported type or is improperly formatted.
-    """
-    value = value.strip().lower()
-    if value in ["str", "text", "string", "varchar"]:
-        return {"type": "string"}
-    elif value in ["int", "integer"]:
-        return {"type": "integer"}
-    elif value in ["float", "decimal", "number"]:
-        return {"type": "number"}
-    elif value in ["bool", "boolean"]:
-        return {"type": "boolean"}
-    elif value.startswith("list["):
-        inner_type = value[5:-1].strip()
-        return {"type": "array", "items": convert_val(inner_type, model)}
-    elif value == "list":
-        raise ValueError("List type must specify its elements, e.g., 'list[str]'")
-    elif value.startswith("{") and value.endswith("}"):
-        # Handle dictionary type
-        properties = {}
-        for item in value[1:-1].split(","):
-            key, val = item.strip().split(":")
-            properties[key.strip()] = convert_val(val.strip(), model)
-        result = {
-            "type": "object",
-            "properties": properties,
-            "required": list(properties.keys()),
-        }
-        # TODO: this is a hack to get around the fact that gemini doesn't support additionalProperties
-        if "gemini" not in model:
-            result["additionalProperties"] = False
-        return result
-    else:
-        raise ValueError(f"Unsupported value type: {value}")
-
-
-def cache_key(
-    model: str,
-    op_type: str,
-    messages: List[Dict[str, str]],
-    output_schema: Dict[str, str],
-    scratchpad: Optional[str] = None,
-) -> str:
-    """
-    Generate a unique cache key based on function arguments.
-
-    This function creates a hash-based key using the input parameters, which can
-    be used for caching purposes.
-
-    Args:
-        model (str): The model name.
-        op_type (str): The operation type.
-        messages (List[Dict[str, str]]): The messages to send to the LLM.
-        output_schema (Dict[str, str]): The output schema dictionary.
-        scratchpad (Optional[str]): The scratchpad to use for the operation.
-
-    Returns:
-        str: A unique hash string representing the cache key.
-    """
-    # Ensure no non-serializable objects are included
-    key_dict = {
-        "model": model,
-        "op_type": op_type,
-        "messages": json.dumps(messages, sort_keys=True),
-        "output_schema": json.dumps(output_schema, sort_keys=True),
-        "scratchpad": scratchpad,
-    }
-    return hashlib.md5(json.dumps(key_dict, sort_keys=True).encode()).hexdigest()
-
-
-def get_user_input_for_schema(schema: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Prompt the user for input for each key in the schema using Rich,
-    then parse the input values with json.loads().
-
-    Args:
-        schema (Dict[str, Any]): The schema dictionary.
-
-    Returns:
-        Dict[str, Any]: A dictionary with user inputs parsed according to the schema.
-    """
-    user_input = {}
-
-    for key, value_type in schema.items():
-        prompt_text = f"Enter value for '{key}' ({value_type}): "
-        user_value = Prompt.ask(prompt_text)
-
-        try:
-            # Parse the input value using json.loads()
-            parsed_value = json.loads(user_value)
-
-            # Check if the parsed value matches the expected type
-            if isinstance(parsed_value, eval(value_type)):
-                user_input[key] = parsed_value
-            else:
-                rprint(
-                    f"[bold red]Error:[/bold red] Input for '{key}' does not match the expected type {value_type}."
-                )
-                return get_user_input_for_schema(schema)  # Recursive call to retry
-
-        except json.JSONDecodeError:
-            rprint(
-                f"[bold red]Error:[/bold red] Invalid JSON input for '{key}'. Please try again."
-            )
-            return get_user_input_for_schema(schema)  # Recursive call to retry
-
-    return user_input
-
-
-class InvalidOutputError(Exception):
-    """
-    Custom exception raised when the LLM output is invalid or cannot be parsed.
-
-    Attributes:
-        message (str): Explanation of the error.
-        output (str): The invalid output that caused the exception.
-        expected_schema (Dict[str, Any]): The expected schema for the output.
-        messages (List[Dict[str, str]]): The messages sent to the LLM.
-        tools (Optional[List[Dict[str, str]]]): The tool calls generated by the LLM.
-    """
-
-    def __init__(
-        self,
-        message: str,
-        output: str,
-        expected_schema: Dict[str, Any],
-        messages: List[Dict[str, str]],
-        tools: Optional[List[Dict[str, str]]] = None,
-    ):
-        self.message = message
-        self.output = output
-        self.expected_schema = expected_schema
-        self.messages = messages
-        self.tools = tools
-        super().__init__(self.message)
-
-    def __str__(self):
-        return (
-            f"{self.message}\n"
-            f"Invalid output: {self.output}\n"
-            f"Expected schema: {self.expected_schema}\n"
-            f"Messages sent to LLM: {self.messages}\n"
-            f"Tool calls generated by LLM: {self.tools}"
-        )
-
-
-def timeout(seconds):
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            result = [TimeoutError("Function call timed out")]
-
-            def target():
-                try:
-                    result[0] = func(*args, **kwargs)
-                except Exception as e:
-                    result[0] = e
-
-            thread = threading.Thread(target=target)
-            thread.start()
-            thread.join(seconds)
-            if isinstance(result[0], Exception):
-                raise result[0]
-            return result[0]
-
-        return wrapper
-
-    return decorator
-
-
-def truncate_messages(
-    messages: List[Dict[str, str]], model: str, from_agent: bool = False
-) -> List[Dict[str, str]]:
-    """
-    Truncate the messages to fit the model's context length.
-    """
-    model_input_context_length = model_cost.get(model.split("/")[-1], {}).get(
-        "max_input_tokens", 8192
-    )
-    total_tokens = sum(count_tokens(json.dumps(msg), model) for msg in messages)
-
-    if total_tokens <= model_input_context_length - 100:
-        return messages
-
-    truncated_messages = messages.copy()
-    longest_message = max(truncated_messages, key=lambda x: len(x["content"]))
-    content = longest_message["content"]
-    excess_tokens = total_tokens - model_input_context_length + 200  # 200 token buffer
-
-    try:
-        encoder = tiktoken.encoding_for_model(model.split("/")[-1])
-    except Exception:
-        encoder = tiktoken.encoding_for_model("gpt-4o")
-    encoded_content = encoder.encode(content)
-    tokens_to_remove = min(len(encoded_content), excess_tokens)
-    mid_point = len(encoded_content) // 2
-    truncated_encoded = (
-        encoded_content[: mid_point - tokens_to_remove // 2]
-        + encoder.encode(f" ... [{tokens_to_remove} tokens truncated] ... ")
-        + encoded_content[mid_point + tokens_to_remove // 2 :]
-    )
-    truncated_content = encoder.decode(truncated_encoded)
-    # Calculate the total number of tokens in the original content
-    total_tokens = len(encoded_content)
-
-    # Print the warning message using rprint
-    warning_type = "User" if not from_agent else "Agent"
-    rprint(
-        f"[yellow]{warning_type} Warning:[/yellow] Cutting {tokens_to_remove} tokens from a prompt with {total_tokens} tokens..."
-    )
-
-    longest_message["content"] = truncated_content
-
-    return truncated_messages
-
-
-def safe_eval(expression: str, output: Dict) -> bool:
-    """
-    Safely evaluate an expression with a given output dictionary.
-    Uses asteval to evaluate the expression.
-    https://lmfit.github.io/asteval/index.html
-    """
-    try:
-        # Add the output dictionary to the symbol table
-        aeval.symtable["output"] = output
-        # Safely evaluate the expression
-        return bool(aeval(expression))
-    except Exception:
-        # try to evaluate with python eval
-        try:
-            return bool(eval(expression, locals={"output": output}))
-        except Exception:
-            return False
+from docetl.utils import completion_cost
+from docetl.helper.cache import cache_key
+from docetl.helper.generic import freezeargs, cache, LLMResult, convert_dict_schema_to_list_schema, truncate_messages, \
+    timeout, convert_val, InvalidOutputError, get_user_input_for_schema, safe_eval
 
 
 class APIWrapper(object):
@@ -423,7 +60,7 @@ class APIWrapper(object):
                 c.set(key, result)
 
         return result
-    
+
     def call_llm_batch(
         self,
         model: str,
@@ -437,10 +74,10 @@ class APIWrapper(object):
     ) -> LLMResult:
         # Turn the output schema into a list of schemas
         output_schema = convert_dict_schema_to_list_schema(output_schema)
-        
+
         # Invoke the LLM call
         return self.call_llm(model, op_type,messages, output_schema, verbose=verbose, timeout_seconds=timeout_seconds, max_retries_per_timeout=max_retries_per_timeout, bypass_cache=bypass_cache)
-        
+
 
     def _cached_call_llm(
         self,
@@ -708,7 +345,7 @@ class APIWrapper(object):
                     initial_result=initial_result,
                 )
             except RateLimitError:
-                # TODO: this is a really hacky way to handle rate limits
+                # TODO:     this is a really hacky way to handle rate limits
                 # we should implement a more robust retry mechanism
                 backoff_time = 4 * (2**rate_limited_attempt)  # Exponential backoff
                 max_backoff = 120  # Maximum backoff time of 60 seconds
@@ -1026,136 +663,3 @@ Remember: The scratchpad should contain information necessary for processing fut
                 console.log(f"[yellow]Output:[/yellow] {output}")
                 return False
         return True
-
-
-class RichLoopBar:
-    """
-    A progress bar class that integrates with Rich console.
-
-    This class provides a wrapper around tqdm to create progress bars that work
-    with Rich console output.
-
-    Args:
-        iterable (Optional[Union[Iterable, range]]): An iterable to track progress.
-        total (Optional[int]): The total number of iterations.
-        desc (Optional[str]): Description to be displayed alongside the progress bar.
-        leave (bool): Whether to leave the progress bar on screen after completion.
-        console: The Rich console object to use for output.
-    """
-
-    def __init__(
-        self,
-        iterable: Optional[Union[Iterable, range]] = None,
-        total: Optional[int] = None,
-        desc: Optional[str] = None,
-        leave: bool = True,
-        console=None,
-    ):
-        if console is None:
-            raise ValueError("Console must be provided")
-        self.console = console
-        self.iterable = iterable
-        self.total = self._get_total(iterable, total)
-        self.description = desc
-        self.leave = leave
-        self.tqdm = None
-
-    def _get_total(self, iterable, total):
-        """
-        Determine the total number of iterations for the progress bar.
-
-        Args:
-            iterable: The iterable to be processed.
-            total: The explicitly specified total, if any.
-
-        Returns:
-            int or None: The total number of iterations, or None if it can't be determined.
-        """
-        if total is not None:
-            return total
-        if isinstance(iterable, range):
-            return len(iterable)
-        try:
-            return len(iterable)
-        except TypeError:
-            return None
-
-    def __iter__(self):
-        """
-        Create and return an iterator with a progress bar.
-
-        Returns:
-            Iterator: An iterator that yields items from the wrapped iterable.
-        """
-        self.tqdm = tqdm(
-            self.iterable,
-            total=self.total,
-            desc=self.description,
-            file=self.console.file,
-        )
-        for item in self.tqdm:
-            yield item
-
-    def __enter__(self):
-        """
-        Enter the context manager, initializing the progress bar.
-
-        Returns:
-            RichLoopBar: The RichLoopBar instance.
-        """
-        self.tqdm = tqdm(
-            total=self.total,
-            desc=self.description,
-            leave=self.leave,
-            file=self.console.file,
-        )
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        """
-        Exit the context manager, closing the progress bar.
-
-        Args:
-            exc_type: The type of the exception that caused the context to be exited.
-            exc_val: The instance of the exception that caused the context to be exited.
-            exc_tb: A traceback object encoding the stack trace.
-        """
-        self.tqdm.close()
-
-    def update(self, n=1):
-        """
-        Update the progress bar.
-
-        Args:
-            n (int): The number of iterations to increment the progress bar by.
-        """
-        if self.tqdm:
-            self.tqdm.update(n)
-
-
-def rich_as_completed(futures, total=None, desc=None, leave=True, console=None):
-    """
-    Yield completed futures with a Rich progress bar.
-
-    This function wraps concurrent.futures.as_completed with a Rich progress bar.
-
-    Args:
-        futures: An iterable of Future objects to monitor.
-        total (Optional[int]): The total number of futures.
-        desc (Optional[str]): Description for the progress bar.
-        leave (bool): Whether to leave the progress bar on screen after completion.
-        console: The Rich console object to use for output.
-
-    Yields:
-        Future: Completed future objects.
-
-    Raises:
-        ValueError: If no console object is provided.
-    """
-    if console is None:
-        raise ValueError("Console must be provided")
-
-    with RichLoopBar(total=total, desc=desc, leave=leave, console=console) as pbar:
-        for future in as_completed(futures):
-            yield future
-            pbar.update()
