+++ tests/basic/test_basic_filter_split_gather.py
@@ -4,8 +4,6 @@ from docetl.operations.unnest import UnnestOperation
 from docetl.operations.equijoin import EquijoinOperation
 from docetl.operations.split import SplitOperation
 from docetl.operations.gather import GatherOperation
-from docetl.operations.utils import APIWrapper
-from docetl.config_wrapper import ConfigWrapper
 from dotenv import load_dotenv
 from tests.conftest import api_wrapper
